refactor: Payment 관련 클래스 및 패키지 구조 리팩토링
