+++ src/main/java/roomescape/payment/PaymentConfirmWebRequest.java
@@ -1,5 +1,7 @@
 package roomescape.payment;
 
+import roomescape.payment.service.dto.PaymentConfirmRequest;
+
 public record PaymentConfirmWebRequest(
         String paymentKey,
         String orderId,
