+++ src/main/java/roomescape/payment/service/PaymentService.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.payment.service;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
@@ -6,6 +6,8 @@ import java.util.Base64.Encoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClient.Builder;
+import roomescape.payment.service.dto.PaymentConfirmRequest;
+import roomescape.payment.service.dto.PaymentConfirmResponse;
 
 @Service
 public class PaymentService {
