+++ src/main/java/roomescape/payment/service/dto/PaymentConfirmRequest.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.payment.service.dto;
 
 public record PaymentConfirmRequest(
         String paymentKey,
