+++ src/main/java/roomescape/payment/service/dto/PaymentConfirmResponse.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.payment.service.dto;
 
 public record PaymentConfirmResponse(
         String paymentKey,
