+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -23,9 +23,9 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.common.exception.PaymentException;
 import roomescape.member.auth.vo.MemberInfo;
 import roomescape.member.domain.Role;
-import roomescape.payment.PaymentConfirmResponse;
+import roomescape.payment.service.dto.PaymentConfirmResponse;
 import roomescape.payment.PaymentConfirmWebRequest;
-import roomescape.payment.PaymentService;
+import roomescape.payment.service.PaymentService;
 import roomescape.reservation.controller.dto.CreateReservationWebRequest;
 import roomescape.reservation.controller.dto.CreateReservationWithPaymentWebRequest;
 import roomescape.reservation.domain.Reservation;
